Split cloud command wrapper
